Moving functionality to background events
Moved a bunch of functions and calls to background.js. New: On install,
it checks for and creates the Dropmarks folder, which can be moved and
renamed. On install, the user key is generated. On startup, pulls any
missing marks from server to local. In progress: pushing mark removal
from local to server.